Add the QBS build script
This is a simple script for the QBS build system. Now it builded
the yaml library as static, but it can also be modified for a variant
of dynamic assembly.
This scenario can be used either for the self-assembly of the static
library, or as part of large QBS project.