meson build system support
Adds a meson.build file to build yaml-cpp. Meson supports downloading
third-party dependencies (like GTest) automatically if they are not
installed. This change allows meson users to use yaml-cpp very easily
as such a fallback, too.

The resulting installation for standard settings is the same as from
CMake but currently, the cmake package information is missing. Instead,
only the pkgconfig file is provided.