CMake: Allow to disable uninstall
add_custom_target uninstall causes a clash with the same target in rtmidi (https://github.com/thestk/rtmidi).
"add_custom_target cannot create target "uninstall" because another target with the same name already exists.  The existing target is a custom target created in source directory"

So it's probably easiest to simply allow to disable the uninstall in the first place.