node/convert: relax the check for string_view
in 35b44980, the conversion for `std::string_view` was added, but
it enables it only if C++17 is used. but the availability of
`std::string_view` does not depends on C++17. in this change,
we use the more specific language feature macro to check the
availability of `std::string_view` instead of checking for the
C++ standard.

Refs #1148
Signed-off-by: Kefu Chai <tchaikov@gmail.com>